Patch the closure type for AbortController.abort() so correct code is generated.
Prior to this patch the AbortController generated a class

@JsType(isNative = true, namespace = JsPackage.GLOBAL)
public class AbortController {
  @JsFunction
  public interface AbortFn {
    Object onInvoke();
  }

  public AbortController.AbortFn abort;
  public AbortSignal signal;
}

However when this was used in a GWT 2.8.2 it would produce incorrect code. For example
_abortController.abort.onInvoke() would be compiled to `this._abortController.abort.call(null)`
which would result in an error from the browser.

The patched closure extern will generate working code that looks like:

@JsType( isNative = true, namespace = JsPackage.GLOBAL )
public class AbortController
{
  public AbortSignal signal;

  public native void abort();
}