Remove values from externs
See #3293 for the reasoning but the short version is that closure
compiler does not use the values and reads the values at runtime.
If any tool were to start using it and the values were incorrect,
then incorrect behaviour will occur.

FWIW, while stripping the values I found that there is at least
one value (XPathException.INVALID_EXPRESSION_ERR) that was incorrectly
defined.

Closes https://github.com/google/closure-compiler/pull/3312.

PiperOrigin-RevId: 244240926