Fix constructor path for Integer/Wildcard conversion.
Using "constructor" as identifier for constructor works for the jsinterop generator because methods come from javascript where constructor is a reserved word. There is no risk of conflict.

In addition to that, this cl renames getJavaFqn name to getConfigurationIdentifier for methods/fields/parameters as fqn doesn't make sense for these entities.

PiperOrigin-RevId: 260598450